+++ projects/ids-enterprise-ng/src/lib/column/soho-column.component.ts
@@ -1,22 +1,26 @@
 /// <reference path="soho-column.d.ts" />
 
 import {
+  AfterViewChecked,
   AfterViewInit,
+  ChangeDetectionStrategy,
   Component,
   ElementRef,
   EventEmitter,
   HostBinding,
   Input,
+  NgZone,
   OnDestroy,
   Output,
 } from '@angular/core';
 
 @Component({
   selector: '[soho-column]', // tslint:disable-line
-  template: '<ng-content></ng-content>'
+  template: '<ng-content></ng-content>',
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 
-export class SohoColumnComponent implements AfterViewInit, OnDestroy {
+export class SohoColumnComponent implements AfterViewInit, AfterViewChecked, OnDestroy {
   /** Options. */
   private options: SohoColumnOptions = {};
 
@@ -30,17 +34,17 @@ export class SohoColumnComponent implements AfterViewInit, OnDestroy {
 
     if (this.column) {
       this.column.settings.dataset = dataset;
-      this.column.updated(this.column.settings);
+      this.updateRequired = true;
     }
   }
 
    /** Chart Type */
-  @Input() set type(value: string) {
+  @Input() set type(value: SohoColumnType) {
     this.options.type = value;
 
     if (this.column) {
       this.column.settings.type = value;
-      this.column.updated(this.column.settings);
+      this.updateRequired = true;
     }
   }
 
@@ -50,7 +54,7 @@ export class SohoColumnComponent implements AfterViewInit, OnDestroy {
 
     if (this.column) {
       this.column.settings.isStacked = value;
-      this.column.updated(this.column.settings);
+      this.updateRequired = true;
     }
   }
 
@@ -60,7 +64,7 @@ export class SohoColumnComponent implements AfterViewInit, OnDestroy {
 
     if (this.column) {
       this.column.settings.showLegend = value;
-      this.column.updated(this.column.settings);
+      this.updateRequired = true;
     }
   }
 
@@ -70,7 +74,7 @@ export class SohoColumnComponent implements AfterViewInit, OnDestroy {
 
     if (this.column) {
       this.column.settings.animate = value;
-      this.column.updated(this.column.settings);
+      this.updateRequired = true;
     }
   }
 
@@ -80,7 +84,7 @@ export class SohoColumnComponent implements AfterViewInit, OnDestroy {
 
     if (this.column) {
       this.column.settings.redrawOnResize = value;
-      this.column.updated(this.column.settings);
+      this.updateRequired = true;
     }
   }
 
@@ -90,7 +94,7 @@ export class SohoColumnComponent implements AfterViewInit, OnDestroy {
 
     if (this.column) {
       this.column.settings.format = value;
-      this.column.updated(this.column.settings);
+      this.updateRequired = true;
     }
   }
 
@@ -99,27 +103,27 @@ export class SohoColumnComponent implements AfterViewInit, OnDestroy {
 
     if (this.column) {
       this.column.settings.formatterString = value;
-      this.column.updated(this.column.settings);
+      this.updateRequired = true;
     }
   }
 
   /** Settings for the chart ticks. Can set ticks: {format: d3Format, number: n} */
-  @Input() set ticks(value: object[]) {
+  @Input() set ticks(value: object) {
     this.options.ticks = value;
 
     if (this.column) {
       this.column.settings.ticks = value;
-      this.column.updated(this.column.settings);
+      this.updateRequired = true;
     }
   }
 
   /** An empty message will be displayed when there is no chart data. */
-  @Input() set emptyMessage(value: object[]) {
+  @Input() set emptyMessage(value: SohoEmptyMessageOptions) {
     this.options.emptyMessage = value;
 
     if (this.column) {
       this.column.settings.emptyMessage = value;
-      this.column.updated(this.column.settings);
+      this.updateRequired = true;
     }
   }
 
@@ -129,7 +133,7 @@ export class SohoColumnComponent implements AfterViewInit, OnDestroy {
 
     if (this.column) {
       this.column.settings.xAxis = value;
-      this.column.updated(this.column.settings);
+      this.updateRequired = true;
     }
   }
 
@@ -139,7 +143,7 @@ export class SohoColumnComponent implements AfterViewInit, OnDestroy {
 
     if (this.column) {
       this.column.settings.yAxis = value;
-      this.column.updated(this.column.settings);
+      this.updateRequired = true;
     }
   }
 
@@ -150,38 +154,61 @@ export class SohoColumnComponent implements AfterViewInit, OnDestroy {
 
   private jQueryElement: JQuery;
   public column: SohoColumn;
-  constructor(private element: ElementRef) { }
+  private updateRequired = false;
+
+  constructor(
+    private element: ElementRef,
+    private ngZone: NgZone,
+  ) { }
 
   /** Setup */
   ngAfterViewInit() {
-    this.jQueryElement = jQuery(this.element.nativeElement);
+    this.ngZone.runOutsideAngular(() => {
+      this.jQueryElement = jQuery(this.element.nativeElement);
 
-    this.jQueryElement.chart(this.options);
-    this.column = this.jQueryElement.data('column');
+      this.jQueryElement.chart(this.options);
+      this.column = this.jQueryElement.data('column');
 
-    // Setup the events
-    this.jQueryElement.on('selected', (e: any, args: SohoColumnSelectEvent) => this.selected.emit(args));
-    this.jQueryElement.on('unselected', (e: any, args: SohoColumnSelectEvent) => this.unselected.emit(args));
-    this.jQueryElement.on('rendered', (...args) => this.rendered.emit(args));
+      // Setup the events
+      this.jQueryElement.on('selected', (e: any, args: SohoColumnSelectEvent) =>
+        this.ngZone.run(() => this.selected.emit(args)));
+      this.jQueryElement.on('unselected', (e: any, args: SohoColumnSelectEvent) =>
+        this.ngZone.run(() => this.unselected.emit(args)));
+      this.jQueryElement.on('rendered', (...args) =>
+        this.ngZone.run(() => this.rendered.emit(args)));
+    });
+  }
+
+  ngAfterViewChecked() {
+    if (this.column && this.updateRequired) {
+      this.ngZone.runOutsideAngular(() => this.column.updated(this.column.settings));
+      this.updateRequired = false;
+    }
   }
 
   /** Tear Down */
   ngOnDestroy() {
-    if (this.column) {
-      this.column.destroy();
-      this.column = null;
-    }
+    // call outside the angular zone so change detection isn't triggered by the soho component.
+    this.ngZone.runOutsideAngular(() => {
+      if (this.jQueryElement) {
+        this.jQueryElement.off();
+      }
+      if (this.column) {
+        this.column.destroy();
+        this.column = null;
+      }
+    });
   }
 
   public setSelected(selected: SohoColumnSelected) {
-    this.column.setSelected(selected);
+    this.ngZone.runOutsideAngular(() => this.column.setSelected(selected));
   }
 
   public toggleSelected(selected: SohoColumnSelected) {
-    this.column.toggleSelected(selected);
+    this.ngZone.runOutsideAngular(() => this.column.toggleSelected(selected));
   }
 
   public getSelected() {
-    this.column.getSelected();
+    return this.ngZone.runOutsideAngular(() => this.column.getSelected());
   }
 }