+++ docs/CHANGELOG.md
@@ -13,6 +13,8 @@
 
 ### 4.12.0 Fixes
 
+- `[General]` Refactored a handful of components to use `ngZone`. This effects the constructor and how often change detection is called. `PWP`
+
 ### 4.12.0 Chore & Maintenance
 
 ## v4.11.0